refactor: move static content
- testimonials (used in multiple places) are in testimonials directory
- rest of specialty page static content is back in page component
