--- content/specialties/next.js
@@ -1,73 +1,0 @@
-export const caseStudies = [
-  {
-    imgAlt:
-      'Enigma hirer dashboard with a pop-up notification introducing the page',
-    imgSrc: '/img/work/enigma/enigma-hirer-dashboard.png',
-    logoSrc: '/svgs/clients/enigma.svg',
-    text: 'Enigma is a platform designed to help cybersecurity teams make faster hiring decisions and find...'
-  }
-];
-
-export const community = {
-  imgSrc: '/img/hero-images/ember-consulting.png',
-  imgHeight: '1034',
-  h2: 'We ❤️ the Next.js Community',
-  paragraph:
-    'We’re a team of makers who are passionate about improving software products and the people who create them.'
-};
-
-export const head = {
-  title: 'Next.js Software Consultants & Developers',
-  description:
-    'Wondering,"What is Next js?" or optimizing your Next application? See how Ship Shape\'s Next.js specialists can help build a universal app without the hassle.',
-  url: 'https://shipshape.io/services/next-consulting/'
-};
-
-export const marketing = {
-  imgSrc: '/img/hero-images/mentoring.png',
-  imgHeight: '772',
-  h2: 'We’ll never leave you high and dry',
-  paragraph:
-    'Ship Shape is committed to leaving every engineering team we work with better than we found it. Not only do we build robust, reliable software, but we also train your staff along the way to understand, maintain, and improve your product after our engagement ends.'
-};
-
-export const specialtyDetails = {
-  text: {
-    h2: 'All about Next.js',
-    h3: 'Some of our favorite Next.js features include:'
-  },
-  benefits: [
-    {
-      imgSrc: '/svgs/icons/image.svg',
-      description: 'Out-of-the-box internationalization features'
-    },
-    {
-      imgSrc: '/svgs/icons/mentor.svg',
-      description:
-        'Automatic image resizing and one-off, on-the-fly image generation'
-    },
-    {
-      imgSrc: '/svgs/icons/chart.svg',
-      description: 'Amazingly high performance'
-    },
-    {
-      imgSrc: '/svgs/icons/pwa.svg',
-      description:
-        'Zero config! Developing features is a simple “yarn dev” away'
-    }
-  ]
-};
-
-export const testimonial = {
-  name: 'Harley Sugarman',
-  title: 'Founder, CEO at Enigma',
-  imgSrc: 'harley',
-  svgSrc: '/svgs/clients/enigma-white.svg',
-  quote: `I loved working with Ship Shape.
-          They helped our small team build a product from the ground up
-          and were strong partners every step of the way. Their domain expertise
-          in Next.js and React helped lay the foundation for a scalable codebase,
-          and they were great at ensuring we adhered to engineering best practices.
-          Every member of the team was a pleasure to work with and (best of all)
-          we ended up launching our MVP on schedule!`
-};
