--- content/specialties/nuxt.js
@@ -1,56 +1,0 @@
-export const community = {
-  imgSrc: '/img/hero-images/ember-consulting.png',
-  imgHeight: '1034',
-  h2: 'We ❤️ the Nuxt.js Community',
-  p: ''
-};
-
-export const head = {
-  title: 'Nuxt.js Software Consultants & Developers',
-  description:
-    "Working to improve or create a Nuxt app? Ship Shape's first-class engineers build fast, lightweight Nuxt.js applications while training your team too.",
-  url: 'https://shipshape.io/services/nuxt-consulting/'
-};
-
-export const marketing = {
-  imgSrc: '/img/hero-images/mentoring.png',
-  imgHeight: '772',
-  h2: "Your first mate's in rough waters",
-  paragraph:
-    'Ship Shape is committed to leaving every engineering team we work with better than we found it. Not only do we build robust, reliable software, but we also train your staff along the way to understand, maintain, and improve your product after our engagement ends.'
-};
-
-export const specialtyDetails = {
-  text: {
-    h2: 'All about Nuxt.js',
-    h3: 'Nuxt works well for apps big and small thanks to its:'
-  },
-  benefits: [
-    {
-      imgSrc: '/svgs/icons/modular.svg',
-      description: 'Modular architecture, which speeds up development'
-    },
-    {
-      imgSrc: '/svgs/icons/native-apps.svg',
-      description:
-        'Server-side rendered (SSR) to deliver HTML, rather than JavaScript'
-    },
-    {
-      imgSrc: '/svgs/icons/static.svg',
-      description:
-        'Static Site Generation to net SEO benefits without needing a server'
-    }
-  ]
-};
-
-export const testimonial = {
-  name: 'Donald Piret',
-  title: 'Director of Engineering at Sephora',
-  imgSrc: 'donald',
-  svgSrc: '/svgs/clients/sephora-white.svg',
-  quote: `Ship Shape delivered beyond our expectations. They pointed out things
-          we had never even thought of or knew about. Our team learned a huge
-          amount of new information and new tricks, and we still have them on
-          retainer for very specific questions. Plus, Ship Shape is the
-          friendliest team of engineers I've worked with.`
-};
