--- content/specialties/serverless.js
@@ -1,52 +1,0 @@
-export const community = {
-  imgSrc: '/img/hero-images/ember-consulting.png',
-  imgHeight: '1034',
-  h2: 'We ❤️ the Developer Community',
-  paragraph:
-    'We’re a team of makers who are passionate about improving software products and the people who create them.'
-};
-
-export const head = {
-  title: 'Serverless Framework Solutions',
-  description:
-    "Turn to Ship Shape's serverless framework experts to better understand what is serverless, and save time and resources on your next app build.",
-  url: 'https://shipshape.io/services/serverless-consulting/'
-};
-
-export const marketing = {
-  imgSrc: '/img/hero-images/mentoring.png',
-  imgHeight: '772',
-  h2: 'Your co-pilots in the storm',
-  paragraph:
-    'Ship Shape is committed to leaving every engineering team we work with better than we found it. Not only do we build robust, reliable software, but we also train your staff along the way to understand, maintain, and improve your product after our engagement ends.'
-};
-
-export const specialtyDetails = {
-  text: {
-    h2: 'All about Serverless',
-    h3: 'This means your company can:'
-  },
-  benefits: [
-    { imgSrc: '/svgs/icons/modular.svg', description: 'Clear Organization' },
-    {
-      imgSrc: '/svgs/icons/chart.svg',
-      description: 'Scalable UI architecture'
-    },
-    { imgSrc: '/svgs/icons/retainer.svg', description: 'Simple updates' },
-    { imgSrc: '/svgs/icons/sail.svg', description: 'Commitment to stability' }
-  ]
-};
-
-export const testimonial = {
-  name: 'Harley Sugarman',
-  title: 'Founder, CEO at Enigma',
-  imgSrc: 'harley',
-  svgSrc: '/svgs/clients/enigma.svg',
-  quote: `I loved working with Ship Shape. 
-          They helped our small team build a product from the ground up 
-          and were strong partners every step of the way. Their domain expertise 
-          in Next.js and React helped lay the foundation for a scalable codebase, 
-          and they were great at ensuring we adhered to engineering best practices. 
-          Every member of the team was a pleasure to work with and (best of all) 
-          we ended up launching our MVP on schedule!`
-};
