--- content/specialties/tailwind.js
@@ -1,63 +1,0 @@
-export const community = {
-  imgSrc: '/img/hero-images/ember-consulting.png',
-  imgHeight: '1034',
-  h2: 'We ❤️ the Developer Community',
-  paragraph:
-    'We’re a team of makers who are passionate about improving software products and the people who create them.'
-};
-
-export const head = {
-  title: 'Tailwind CSS & Tailwind UI Experts',
-  description:
-    "Tailwind CSS makes it easy to build a site without using any custom CSS. Learn how Ship Shape's team of Tailwind UI experts can help tailor your app quickly. ",
-  url: 'https://shipshape.io/services/tailwind-consulting/'
-};
-
-export const marketing = {
-  imgSrc: '/img/hero-images/mentoring.png',
-  imgHeight: '772',
-  h2: 'Your co-pilots in the storm',
-  paragraph:
-    'Ship Shape is committed to leaving every engineering team we work with better than we found it. Not only do we build robust, reliable software, but we also train your staff along the way to understand, maintain, and improve your product after our engagement ends.'
-};
-
-export const specialtyDetails = {
-  text: {
-    h2: 'All about Tailwind CSS',
-    h3: 'Ship Shape’s experienced engineers love Tailwind CSS because:'
-  },
-  benefits: [
-    {
-      imgSrc: '/svgs/icons/image.svg',
-      description: 'It matches mocks and designs with ease.'
-    },
-    {
-      imgSrc: '/svgs/icons/time.svg',
-      description:
-        'It automatically removes all unused CSS when building for production, which improves speed.'
-    },
-    {
-      imgSrc: '/svgs/icons/modular.svg',
-      description:
-        'Its usability is quite nimble compared to similar frameworks (like Bootstrap).'
-    },
-    {
-      imgSrc: '/svgs/icons/pwa.svg',
-      description: 'Prototyping is super fast with Tailwind UI components.'
-    }
-  ]
-};
-
-export const testimonial = {
-  name: 'Harley Sugarman',
-  title: 'Founder, CEO at Enigma',
-  imgSrc: 'harley',
-  svgSrc: '/svgs/clients/enigma.svg',
-  quote: `I loved working with Ship Shape.
-          They helped our small team build a product from the ground up
-          and were strong partners every step of the way. Their domain expertise
-          in Next.js and React helped lay the foundation for a scalable codebase,
-          and they were great at ensuring we adhered to engineering best practices.
-          Every member of the team was a pleasure to work with and (best of all)
-          we ended up launching our MVP on schedule!`
-};
