+++ content/testimonials/index.js
@@ -1,0 +1,39 @@
+export const enigmaTestimonial = {
+  name: 'Harley Sugarman',
+  title: 'Founder, CEO at Enigma',
+  imgSrc: 'harley',
+  svgSrc: '/svgs/clients/enigma-white.svg',
+  quote: `I loved working with Ship Shape.
+          They helped our small team build a product from the ground up
+          and were strong partners every step of the way. Their domain expertise
+          in Next.js and React helped lay the foundation for a scalable codebase,
+          and they were great at ensuring we adhered to engineering best practices.
+          Every member of the team was a pleasure to work with and (best of all)
+          we ended up launching our MVP on schedule!`
+};
+
+export const expelTestimonial = {
+  name: 'Roger Studner',
+  title: 'Chief Architect at Expel',
+  imgSrc: 'roger',
+  svgSrc: '/svgs/clients/expel-white.svg',
+  quote: `Ship Shape was a great aid to us because they were the voice of Ember
+        best practices to keep our code base current, and they were able to
+        deliver on critical features on our roadmap that we did not have the
+        bandwidth to complete on time. I would absolutely recommend Ship Shape
+        to any other business that is looking for an incredibly skilled
+        partner to work closely with their teams to produce and deliver the
+        highest value.`
+};
+
+export const sephoraTestimonial = {
+  name: 'Donald Piret',
+  title: 'Director of Engineering at Sephora',
+  imgSrc: 'donald',
+  svgSrc: '/svgs/clients/sephora-white.svg',
+  quote: `Ship Shape delivered beyond our expectations. They pointed out things
+          we had never even thought of or knew about. Our team learned a huge
+          amount of new information and new tricks, and we still have them on
+          retainer for very specific questions. Plus, Ship Shape is the
+          friendliest team of engineers I've worked with.`
+};
