+++ pages/index.vue
@@ -43,33 +43,10 @@
 
 <script>
 import { generateMeta } from '~/utils/meta';
-
-const testimonials = [
-  {
-    name: 'Donald Piret',
-    title: 'Director of Engineering at Sephora',
-    imgSrc: 'donald',
-    svgSrc: '/svgs/clients/sephora-white.svg',
-    quote: `Ship Shape delivered beyond our expectations. They pointed out things
-          we had never even thought of or knew about. Our team learned a huge
-          amount of new information and new tricks, and we still have them on
-          retainer for very specific questions. Plus, Ship Shape is the
-          friendliest team of engineers I've worked with.`
-  },
-  {
-    name: 'Roger Studner',
-    title: 'Chief Architect at Expel',
-    imgSrc: 'roger',
-    svgSrc: '/svgs/clients/expel-white.svg',
-    quote: `Ship Shape was a great aid to us because they were the voice of Ember
-          best practices to keep our code base current, and they were able to
-          deliver on critical features on our roadmap that we did not have the
-          bandwidth to complete on time. I would absolutely recommend Ship Shape
-          to any other business that is looking for an incredibly skilled
-          partner to work closely with their teams to produce and deliver the
-          highest value.`
-  }
-];
+import {
+  sephoraTestimonial,
+  expelTestimonial
+} from '~/content/testimonials/index.js';
 
 export default {
   async asyncData({ $content }) {
@@ -78,6 +55,7 @@ export default {
       .limit(3)
       .fetch();
 
+    const testimonials = [sephoraTestimonial, expelTestimonial];
     return { posts, testimonials };
   },
   head() {
