+++ pages/services/ember-consulting.vue
@@ -69,13 +69,60 @@
 
 <script>
 import { generateMeta } from '~/utils/meta';
-import {
-  caseStudies,
-  head,
-  marketing,
-  specialtyDetails,
-  testimonial
-} from '~/content/specialties/ember.js';
+import { expelTestimonial as testimonial } from '~/content/testimonials/index.js';
+
+const caseStudies = [
+  {
+    imgAlt:
+      'An overview of NIST scores, broken down by category and visualized in a radar graph.',
+    imgSrc: '/img/work/expel/nist-overview.png',
+    logoSrc: '/svgs/clients/expel.svg',
+    text: 'Expel offers 24x7 security monitoring and response for cloud, hybrid and on-premises environments...'
+  },
+  {
+    imgAlt: '',
+    imgSrc: '/svgs/data-graphic.svg',
+    logoSrc: '/svgs/clients/netflix.svg',
+    text: "Netflix is the world's leading internet entertainment service with more than 130 million memberships..."
+  },
+  {
+    imgAlt: 'Jebbit Experience Builder Legacy',
+    imgSrc: '/img/work/jebbit/jebbit-case-study-screenshot.png',
+    logoSrc: '/svgs/clients/jebbit.svg',
+    text: 'Jebbit is an enterprise software platform that enables a future of total data transparency...'
+  }
+];
+
+const head = {
+  title: 'Ember.js Enterprise App Development',
+  description:
+    "Learn how Ship Shape's top-tier enterprise app development experts can help your team build fast, reliable Ember.js applications.",
+  url: 'https://shipshape.io/services/ember-consulting/'
+};
+
+const marketing = {
+  imgSrc: '/img/hero-images/mentoring.png',
+  imgHeight: '772',
+  h2: 'We’ll never leave you high and dry',
+  paragraph:
+    'Ship Shape is committed to leaving every engineering team we work with better than we found it. Not only do we build robust, reliable software, but we also train your staff along the way to understand, maintain, and improve your product after our engagement ends.'
+};
+
+const specialtyDetails = {
+  text: {
+    h2: 'All about Ember.js',
+    h3: 'Ember is great for large companies because of its:'
+  },
+  benefits: [
+    { imgSrc: '/svgs/icons/modular.svg', description: 'Clear Organization' },
+    {
+      imgSrc: '/svgs/icons/chart.svg',
+      description: 'Scalable UI architecture'
+    },
+    { imgSrc: '/svgs/icons/retainer.svg', description: 'Simple updates' },
+    { imgSrc: '/svgs/icons/sail.svg', description: 'Commitment to stability' }
+  ]
+};
 
 export default {
   speedkitComponents: {
@@ -87,11 +134,9 @@ export default {
     SingleQuote: () => import('@/components/SingleQuote'),
     WaveMoreLinks: () => import('@/components/WaveMoreLinks')
   },
-
   data() {
     return { caseStudies, head, marketing, testimonial, specialtyDetails };
   },
-
   head() {
     const { title, description, url } = head;
